# addons/swagger.yaml
openapi: 3.0.0
info:
  title: LAB_AP
  version: 1.0.0
paths:
  /events:
    get:
      summary: Get all events
      
      tags:
        - Events
      responses:
        '200':
          description: A list of events
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '500':
          description: Server error
        '404':
          description: Page not found
    post:
      summary: Post an event
      tags:
        - Events
      parameters:
        - name: EventName
          in: query
          description: Event name
          required: true
          schema:
            type: string
        - name: EventDescription
          in: query
          description: Description of the event
          required: true
          schema:
            type: string
        - name: TotalTickets
          in: query
          description: Number of tickets
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Successfully created event!
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '403':
          description: Forbidden
        '401':
          description: Unauthorized
        '500':
          description: Server error
        '400':
          description: This event is already exists

  /events/{event_id}:
    get:
      summary: Get details of a specific event
      tags:
        - Events
      parameters:
        - name: event_id
          in: path
          description: ID of the event
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Event details retrieved successfully
        '400':
          description: Bad request, event with this name is already exist.
        '404':
          description: Event not found
    delete:
      summary: Delete an event
      tags:
        - Events
      parameters:
        - name: event_id
          in: path
          description: ID of the event
          required: true
          schema:
            type: integer
            
      responses:
        '200':
          description: Event successfully deleted
        '400':
          description: Bad request, you dont have permission to delete this event
        '404':
          description: Event not found
    put:
      summary: Update event details
      tags:
        - Events
      parameters:
        - name: event_id
          in: path
          description: ID of the event to update
          required: true
          schema:
            type: integer
      requestBody:
        description: Event details to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventDTO'
      responses:
        '200':
          description: Event successfully updated
        '401':
          description: Unauthorized user
        '400':
          description: Bad request, you dont have permission to update this event
        '404':
          description: Event not found

#Users

  /dashboard:
    get:
      summary: Profile page
      tags:
        - User
      parameters:
        - name: user_id
          in: query
          description: IDs of the events
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Information about user
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserTickets'
        '404':
          description: User not found
        '500':
          description: Server error
          
  /buy-tickets:
    post:
      summary: Buy tickets
      tags:
        - User
      parameters:
        - name: tickets
          in: query
          description: IDs of the events
          required: true
          schema:
            type: array
            items:
              type: integer
      responses:
        '302':
          description: Going to payment page
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserTickets'
        '401':
          description: Unauthorized
        '400':
          description: Bad request
        '404':
          description: User not found
        '500':
          description: Server error
          
  /payment:
    post:
      summary: Payment for tickets
      tags:
        - User
      parameters:
        - name: tickets
          in: query
          description: IDs of the events
          required: true
          schema:
            type: array
            items:
              type: integer
      responses:
        '200':
          description: Successful payment
        '401':
          description: Unauthorized
        '400':
          description: Bad request
        '404':
          description: Some item not found 
        '500':
          description: Server error

  /login:
    post:
      summary: Delete an event
      tags:
        - User
      parameters:
        - name: username
          in: query
          description: username
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: password
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful login
        '400':
          description: Bad request
        '404':
          description: User not found
        '500':
          description: Server error

  /logout:
    get:
      summary: Log out
      tags:
        - User
      responses:
        '200':
          description: Successful logout
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /register:
    post:
      summary: Register user
      tags:
        - User
      parameters:
        - name: username
          in: query
          description: username
          required: true
          schema:
            type: string
        - name: email
          in: query
          description: email
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: password
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful register
        '401':
          description: Unauthorized
        '400':
          description: User already exists
        '500':
          description: Server error

#Ticket

  /event/{event_id}:
    get:
      summary: Get all tickets for an event
      tags:
        - Ticket
      parameters:
        - name: event_id
          in: path
          description: ID of the event
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of tickets
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '500':
          description: Server error
    delete:
      summary: Delete all tickets for event
      tags:
        - Ticket
      parameters:
        - name: event_id
          in: path
          description: ID of the event
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted tickets
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '404':
          description: Ticket not found
        '403':
          description: Forbidden
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /event/{event_id}/tickets/{ticket_id}:
    get:
      summary: Get a specific ticket info
      tags:
        - Ticket
      parameters:
        - name: event_id
          in: path
          description: ID of the event
          required: true
          schema:
            type: integer
        - name: ticket_id
          in: path
          description: ID of the ticket
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted ticket
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '404':
          description: Ticket not found
        '500':
          description: Server error
    delete:
      summary: Delete a specific ticket
      tags:
        - Ticket
      parameters:
        - name: event_id
          in: path
          description: ID of the event
          required: true
          schema:
            type: integer
        - name: ticket_id
          in: path
          description: ID of the ticket
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted ticket
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '404':
          description: Ticket not found
        '403':
          description: Forbidden
        '401':
          description: Unauthorized
        '500':
          description: Server error

#Schemas

components:
  schemas:
  
    Event:
      type: object
      properties:
        Event_ID:
          type: integer
        EventName:
          type: string
        TotalTickets:
          type: string
        EventDescription:
          type: string

    Ticket:
      type: object
      properties:
        Ticket_ID:
          type: integer
        User_ID:
          type: integer
        Event_ID:
          type: integer
        Status: 
          type: string

    User:
      type: object
      properties:
        User_ID:
          type: integer
        Username:
          type: string
        Email:
          type: string
        Password: 
          type: string
    
    UserTickets:
      type: object
      properties:
        Event_ID:
          type: integer
        EventName:
          type: string
        Ticket_ID:
          type: integer
        TicketStatus:
          type: string
          
    EventDTO:
      type: object
      properties:
        EventName:
          type: string
        TotalTickets:
          type: string
        EventDescription:
          type: string






